Test Movie Booking Website w REST Endpoints - CS 4300 Advanced Software Engineering/movie_theater_booking/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-bootstrap5==23.4
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
sqlparse==0.5.1
